requirements.txt
alembic==1.15.1
aniso8601==10.0.0
attrs==25.2.0
bcrypt==4.3.0
blinker==1.9.0
click==8.1.8
Flask==3.1.0
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.7.1
Flask-Mail==0.10.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
Mako==1.3.9
MarkupSafe==3.0.2
PyJWT==2.10.1
python-dotenv==1.0.1
python-slugify==8.0.4
pytz==2025.1
referencing==0.36.2
rpds-py==0.23.1
SQLAlchemy==2.0.38
text-unidecode==1.3
typing_extensions==4.12.2
Werkzeug==3.1.3
